requirements.txt

cupy-cuda12x>=12.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
Pillow>=10.0.0
psutil>=5.9.0
requests>=2.31.0